Refactor markdown api to decouple from vscode extension
